move munit related files to a specific folder
