+++ test-framework/munit/src/main/scala/com/dimafeng/testcontainers/munit/TestContainersForEach.scala
@@ -1,14 +1,14 @@
-package com.dimafeng.testcontainers
+package com.dimafeng.testcontainers.munit
 
 import com.dimafeng.testcontainers.lifecycle.Andable
 import munit.Suite
 
 /**
- * Starts containers before all tests and stop then after all tests
+ * Starts containers before each test and stop them after each test
  *
  * Example:
  * {{{
- * class ExampleSpec extends FunSuite with TestContainersForAll {
+ * class ExampleSpec extends FunSuite with TestContainersForEach {
  *
  *   // First of all, you need to declare, which containers you want to use
  *   override type Containers = MySQLContainer and PostgreSQLContainer
@@ -32,39 +32,28 @@ import munit.Suite
  * }
  *
  * Notes:
- * - If you override beforeAll() without calling super.beforeAll() your containers won't start
- * - If you override afterAll() without calling super.afterAll() your containers won't stop
+ * - If you override beforeEach() without calling super.beforeEach() your containers won't start
+ * - If you override afterEach() without calling super.afterEach() your containers won't stop
  * }}}
  */
-trait TestContainersForAll extends TestContainersSuite { self: Suite =>
+trait TestContainersForEach extends TestContainersSuite { self: Suite =>
   type Containers <: Andable
 
-  override def beforeAll(): Unit = {
-    val tests = self.munitTests()
-    val allTestsIgnored = tests.forall(_.tags.contains(munit.Ignore))
-
-    if (tests.nonEmpty && !allTestsIgnored) {
-      val containers = startContainers()
-      startedContainers = Some(containers)
-      try {
-        afterContainersStart(containers)
-      } catch {
-        case e: Throwable =>
-          stopContainers(containers)
-          throw e
-      }
-    }
-  }
-
   override def beforeEach(context: BeforeEach): Unit = {
-    startedContainers.foreach(beforeTest)
+    val containers = startContainers()
+    startedContainers = Some(containers)
+    try {
+      afterContainersStart(containers)
+      beforeTest(containers)
+    } catch {
+      case e: Throwable =>
+        stopContainers(containers)
+        throw e
+    }
   }
 
   override def afterEach(context: AfterEach): Unit = {
     startedContainers.foreach(afterTest(_, None)) // TODO there is no way to retrieve test status in MUnit - https://github.com/scalameta/munit/issues/119
-  }
-
-  override def afterAll(): Unit = {
     startedContainers.foreach(stopContainers)
   }
 }
