+++ test-framework/munit/src/test/scala/com/dimafeng/testcontainers/munit/TestContainerForEachSpec.scala
@@ -1,15 +1,16 @@
-package com.dimafeng.testcontainers
+package com.dimafeng.testcontainers.munit
 
 import java.util.Optional
 
+import com.dimafeng.testcontainers.ContainerDef
+import com.dimafeng.testcontainers.munit.SampleContainer.SampleJavaContainer
+import com.dimafeng.testcontainers.munit.TestContainerForEachSpec._
 import munit.{FunSuite, MUnitRunner, Suite}
 import org.junit.runner.notification.RunNotifier
 import org.mockito.ArgumentMatchers.any
-import org.mockito.{ArgumentMatchers, Mockito}
 import org.mockito.Mockito.{times, verify}
+import org.mockito.{ArgumentMatchers, Mockito}
 import org.scalatestplus.mockito.MockitoSugar
-import TestContainerForEachSpec._
-import com.dimafeng.testcontainers.SampleContainer.SampleJavaContainer
 
 class TestContainerForEachSpec extends FunSuite with MockitoSugar {
   test("call all appropriate methods of the containers") {
