+++ test-framework/munit/src/test/scala/com/dimafeng/testcontainers/munit/TestContainersForAllSpec.scala
@@ -1,18 +1,18 @@
-package com.dimafeng.testcontainers
+package com.dimafeng.testcontainers.munit
 
 import java.util.Optional
 
-import com.dimafeng.testcontainers.SampleContainer.SampleJavaContainer
-import com.dimafeng.testcontainers.TestContainersForEachSpec._
 import com.dimafeng.testcontainers.lifecycle.and
+import com.dimafeng.testcontainers.munit.SampleContainer.SampleJavaContainer
+import com.dimafeng.testcontainers.munit.TestContainersForAllSpec._
 import munit.{FunSuite, MUnitRunner, Suite}
 import org.junit.runner.notification.RunNotifier
-import org.mockito.{ArgumentMatchers, Mockito}
 import org.mockito.ArgumentMatchers.any
 import org.mockito.Mockito.{times, verify}
+import org.mockito.{ArgumentMatchers, Mockito}
 import org.scalatestplus.mockito.MockitoSugar
 
-class TestContainersForEachSpec extends FunSuite with MockitoSugar {
+class TestContainersForAllSpec extends FunSuite with MockitoSugar {
   test("call all appropriate methods of the containers") {
     val container1 = mock[SampleJavaContainer]
     val container2 = mock[SampleJavaContainer]
@@ -36,7 +36,6 @@ class TestContainersForEachSpec extends FunSuite with MockitoSugar {
     val container2 = mock[SampleJavaContainer]
 
     val spec = new TestSpec(assert(cond = false), container1, container2)
-
     run(spec)
 
     verify(container1).beforeTest(any())
@@ -50,7 +49,7 @@ class TestContainersForEachSpec extends FunSuite with MockitoSugar {
     verify(container2).stop()
   }
 
-  test("start and stop containers before and after each test case") {
+  test("start and stop containers only once") {
     val container1 = mock[SampleJavaContainer]
     val container2 = mock[SampleJavaContainer]
 
@@ -58,14 +57,14 @@ class TestContainersForEachSpec extends FunSuite with MockitoSugar {
     run(spec)
 
     verify(container1, times(2)).beforeTest(any())
-    verify(container1, times(2)).start()
+    verify(container1).start()
     verify(container1, times(2)).afterTest(any(), any())
-    verify(container1, times(2)).stop()
+    verify(container1).stop()
 
     verify(container2, times(2)).beforeTest(any())
-    verify(container2, times(2)).start()
+    verify(container2).start()
     verify(container2, times(2)).afterTest(any(), any())
-    verify(container2, times(2)).stop()
+    verify(container2).stop()
   }
 
   test("call afterContainersStart() and beforeContainersStop()") {
@@ -76,7 +75,7 @@ class TestContainersForEachSpec extends FunSuite with MockitoSugar {
     @volatile var afterStartCalled = false
     @volatile var beforeStopCalled = false
 
-    val spec = new MultipleTestsSpec(assert(cond = true), container1, container2) {
+    val spec = new MultipleTestsSpec(assert(true), container1, container2) {
       override def afterContainersStart(containers: Containers): Unit = {
         super.afterContainersStart(containers)
         afterStartCalled = true
@@ -93,14 +92,14 @@ class TestContainersForEachSpec extends FunSuite with MockitoSugar {
     assert(afterStartCalled && beforeStopCalled)
   }
 
-  test("call beforeContainersStop() and stop container if error is thrown in afterContainersStart()") {
+  test("call beforeContainersStop() and stop container if error thrown in afterContainersStart()") {
     val container1 = mock[SampleJavaContainer]
     val container2 = mock[SampleJavaContainer]
 
     @volatile var afterStartCalled = false
     @volatile var beforeStopCalled = false
 
-    val spec = new MultipleTestsSpec(assert(cond = true), container1, container2) {
+    val spec: MultipleTestsSpec = new MultipleTestsSpec(assert(cond = true), container1, container2) {
       override def afterContainersStart(containers: Containers): Unit = {
         afterStartCalled = true
         throw new RuntimeException("Test")
@@ -115,14 +114,14 @@ class TestContainersForEachSpec extends FunSuite with MockitoSugar {
     run(spec)
 
     verify(container1, times(0)).beforeTest(any())
-    verify(container1, times(2)).start()
+    verify(container1).start()
     verify(container1, times(0)).afterTest(any(), any())
-    verify(container1, times(2)).stop()
+    verify(container1).stop()
 
     verify(container2, times(0)).beforeTest(any())
-    verify(container2, times(2)).start()
+    verify(container2).start()
     verify(container2, times(0)).afterTest(any(), any())
-    verify(container2, times(2)).stop()
+    verify(container2).stop()
 
     assert(afterStartCalled && beforeStopCalled)
   }
@@ -138,10 +137,22 @@ class TestContainersForEachSpec extends FunSuite with MockitoSugar {
 
     verify(container1, Mockito.never()).start()
     verify(container2, Mockito.never()).start()
+
     assert(!called)
   }
 
-  test("not start container if beforeEach is overridden without calling super.beforeEach()") {
+  test("not start container for empty suite") {
+    val container1 = mock[SampleJavaContainer]
+    val container2 = mock[SampleJavaContainer]
+
+    val spec = new EmptySpec(container1, container2)
+    run(spec)
+
+    verify(container1, Mockito.never()).start()
+    verify(container2, Mockito.never()).start()
+  }
+
+  test("not start container if beforeAll is overridden without calling super.beforeAll()") {
     val container1 = mock[SampleJavaContainer]
     val container2 = mock[SampleJavaContainer]
 
@@ -159,7 +170,7 @@ class TestContainersForEachSpec extends FunSuite with MockitoSugar {
     verify(container2, Mockito.never()).stop()
   }
 
-  test("not stop container if afterAll is overridden without calling super.beforeEach()") {
+  test("not stop container if afterAll is overridden without calling super.afterAll()") {
     val container1 = mock[SampleJavaContainer]
     val container2 = mock[SampleJavaContainer]
 
@@ -168,32 +179,21 @@ class TestContainersForEachSpec extends FunSuite with MockitoSugar {
 
     verify(container1).beforeTest(any())
     verify(container1).start()
-    verify(container1, Mockito.never()).afterTest(any(), ArgumentMatchers.eq(Optional.empty()))
+    verify(container1).afterTest(any(), ArgumentMatchers.eq(Optional.empty()))
     verify(container1, Mockito.never()).stop()
 
     verify(container2).beforeTest(any())
     verify(container2).start()
-    verify(container2, Mockito.never()).afterTest(any(), ArgumentMatchers.eq(Optional.empty()))
+    verify(container2).afterTest(any(), ArgumentMatchers.eq(Optional.empty()))
     verify(container2, Mockito.never()).stop()
-  }
-
-  test("not start container for empty suite") {
-    val container1 = mock[SampleJavaContainer]
-    val container2 = mock[SampleJavaContainer]
-
-    val spec = new EmptySpec(container1, container2)
-    run(spec)
-
-    verify(container1, Mockito.never()).start()
-    verify(container2, Mockito.never()).start()
   }
 }
 
-object TestContainersForEachSpec {
+object TestContainersForAllSpec {
   protected abstract class ContainersSpec(
     container1: SampleJavaContainer,
     container2: SampleJavaContainer
-  ) extends FunSuite with TestContainersForEach {
+  ) extends FunSuite with TestContainersForAll {
     override type Containers = SampleContainer and SampleContainer
 
     override def startContainers(): Containers = {
@@ -250,7 +250,7 @@ object TestContainersForEachSpec {
   protected class TestSpecBeforeOverride(testBody: => Unit, container1: SampleJavaContainer, container2: SampleJavaContainer)
     extends ContainersSpec(container1, container2) {
 
-    override def beforeEach(context: BeforeEach): Unit = {}
+    override def beforeAll(): Unit = {}
 
     test("test") {
       withContainers { _ => testBody }
@@ -260,7 +260,7 @@ object TestContainersForEachSpec {
   protected class TestSpecAfterOverride(testBody: => Unit, container1: SampleJavaContainer, container2: SampleJavaContainer)
     extends ContainersSpec(container1, container2) {
 
-    override def afterEach(context: AfterEach): Unit = {}
+    override def afterAll(): Unit = {}
 
     test("test") {
       withContainers { _ => testBody }
